Using cell for Mailer body
