+++ lib/tyrant/config/error_messages.yml
@@ -3,7 +3,9 @@ en:
     user_exists?: "User not found"
     password_ok?: "Wrong Password"
     must_be_equal?: "Passwords are not matching"
-    new_must_match?: "The New Password is not matching"
+    new_must_match?: "Passwords are not matching"
     new_password_must_be_new?: "New password can't match the old one"
     unique_email?: "This email has been already used"
     email?: "Wrong format"
+    safe_url_ok?: "Something went wrong please try to reset the password again"
+    safe_url_expired?: "Link expired"
