+++ lib/tyrant/mailer/view/reset_email.slim
@@ -1,0 +1,13 @@
+h1
+  | Password Reset
+
+P
+  | To reset your password click the link below:
+p
+  = link
+p
+  | This link will expire in 1 hour.
+p
+  | If you did not request your password to be reset, just ignore this email and your password will continue to stay the same,
+p
+  | Cheers
