+++ test/test_helper.rb
@@ -37,3 +37,6 @@ end
 # load features
 
 require "tyrant/sign_up"
+require "tyrant/reset_password"
+require "tyrant/operation/change_password.rb"
+
