+++ tyrant.gemspec
@@ -37,7 +37,6 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "activerecord"
 
-  # spec.add_dependency "trailblazer-cells"
-  # spec.add_dependency 'cells-rails'
-  # spec.add_dependency 'cells-slim'
+  spec.add_dependency 'cells'
+  spec.add_dependency 'cells-slim'
 end
